fix: tipo declarado no calculo de data do job de prazo
"operator is not unique: date + unknown". `CURRENT_DATE + $1` e ambiguo: o
Postgres tem `date + integer` (dias) e `date + interval`, e com o
parametro sem tipo nao escolhe.

SEGUNDO DEFEITO IGUAL NO MESMO DIA. De manha foi "inconsistent types
deduced for parameter $2" no INSERT do pedido de vaga. Mesma causa:
parametro sem tipo declarado. A regra passa a ser: todo parametro que
entra em conta aritmetica leva o tipo escrito.

O teste agora olha o TEXTO do comando, e nao so o resultado. O dublê nao e
Postgres e aceita qualquer SQL — foi por isso que os dois defeitos
passaram por 200 testes e so apareceram no banco de verdade. Ha uma
asserção que falha se sobrar QUALQUER parametro sem tipo nessa consulta.

+++ job_notificacoes.js
@@ -44,6 +44,12 @@ const LIMITE = valorArg('--limite', 500);
  * recalculo nada aqui, e não escrevo em `prestacoes_contas`: este job só lê de lá.
  */
 async function buscarAlvos(pool) {
+  // ⚠️ Os `::int` NÃO são enfeite — sem eles isto morre com
+  // "operator is not unique: date + unknown". `CURRENT_DATE + $1` é ambíguo: o Postgres tem
+  // `date + integer` (dias) e `date + interval`, e com o parâmetro sem tipo não sabe qual
+  // usar. Aconteceu em produção em 10/08, e é o mesmo defeito do INSERT de mais cedo:
+  // parâmetro sem tipo declarado. A regra aqui virou: todo parâmetro em conta aritmética
+  // leva o tipo escrito.
   const { rows } = await pool.query(
     `SELECT codigo_pc, tr, entidade, analista_id, setorial_id, dt_limite_pc,
             (dt_limite_pc::date - CURRENT_DATE) AS dias
@@ -51,9 +57,9 @@ async function buscarAlvos(pool) {
       WHERE analista_id IS NOT NULL
         AND baixada = false
         AND dt_limite_pc IS NOT NULL
-        AND dt_limite_pc::date <= CURRENT_DATE + $1
+        AND dt_limite_pc::date <= CURRENT_DATE + $1::int
       ORDER BY dt_limite_pc
-      LIMIT $2`, [DIAS_AVISO_PREVIO, LIMITE]);
+      LIMIT $2::int`, [DIAS_AVISO_PREVIO, LIMITE]);
   return rows;
 }
